requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
jinja2==3.1.3
starlette==0.36.3
python-multipart==0.0.9
httpx==0.26.0
pydantic>=2.6
python-dotenv>=1.0
google-genai>=0.3
requests>=2.31
pyswisseph>=2.10
bcrypt==4.0.1
sqlalchemy>=2.0
passlib[bcrypt]==1.7.4
itsdangerous>=2.1
stripe>=12.0
PyYAML>=6.0

psycopg[binary]>=3.1
